end step5 - possible issues

+++ src/main/java/com/users/beans/UserProperty.java
@@ -14,14 +14,13 @@ public class UserProperty {
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 
-	private long userId;
+	private int userId;
 	private String propName;
 	private String propValue;
 
-	protected UserProperty() {
-	}
+	protected UserProperty() {}
 
-	public UserProperty(long userId, String propName, String propValue) {
+	public UserProperty(int userId, String propName, String propValue) {
 		this.userId = userId;
 		this.propName = propName;
 		this.propValue = propValue;
@@ -29,11 +28,10 @@ public class UserProperty {
 
 	@Override
 	public String toString() {
-		return "UserProperty [id=" + id + ", userId=" + userId + ", propName=" + propName + ", propValue=" + propValue
-				+ "]";
+		return "UserProperty [id=" + id + ", userId=" + userId + ", propName=" + propName + ", propValue=" + propValue + "]";
 	}
 
-	public long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
